+++ src/main/java/com/example/pharmacymanagementsystem_qlht/controller/CN_TimKiem/ChiTietHoaDon_Ctrl.java
@@ -1,4 +1,4 @@
-package com.example.pharmacymanagementsystem_qlht.controller;
+package com.example.pharmacymanagementsystem_qlht.controller.CN_TimKiem;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -9,7 +9,7 @@ import javafx.stage.Stage;
 public class ChiTietHoaDon_Ctrl extends Application {
     @Override
     public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("/com/example/pharmacymanagementsystem_qlht/ChiTietHoaDon_GUI.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("/com/example/pharmacymanagementsystem_qlht/CN_TimKiem/TKHoaDon/ChiTietHoaDon_GUI.fxml"));
         Scene scene = new Scene(root);
 
         stage.setScene(scene);
