+++ src/main/java/com/example/pharmacymanagementsystem_qlht/controller/CN_TimKiem/TimKiemHoaDon_Ctrl.java
@@ -9,7 +9,7 @@ import javafx.stage.Stage;
 public class TimKiemHoaDon_Ctrl extends Application {
     @Override
     public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("/com/example/pharmacymanagementsystem_qlht/CN_TimKiem/TKHoaDon_GUI.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("/com/example/pharmacymanagementsystem_qlht/CN_TimKiem/TKHoaDon/TKHoaDon_GUI.fxml"));
         Scene scene = new Scene(root);
         scene.getStylesheets().add(getClass().getResource("/com/example/pharmacymanagementsystem_qlht/css/TimKiemThuoc.css").toExternalForm());
         stage.setScene(scene);
