--- src/main/java/com/example/pharmacymanagementsystem_qlht/controller/LapHoaDon_Ctrl.java
@@ -1,20 +1,0 @@
-package com.example.pharmacymanagementsystem_qlht.controller;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class LapHoaDon_Ctrl extends Application {
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        FXMLLoader loader = new FXMLLoader(getClass().getResource("/com/example/pharmacymanagementsystem_qlht/CN_BanHang/CN_LapHoaDon/LapHoaDon_GUI.fxml"));
-        Parent root = loader.load();
-        Scene scene = new Scene(root);
-        stage.setScene(scene);
-        stage.show();
-    }
-
-}
