+++ src/main/resources/com/example/pharmacymanagementsystem_qlht/CN_QuanLy/QLKeHang/ThemKeHang_GUI.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="com.example.pharmacymanagementsystem_qlht.controller.CN_QuanLy.QLKeHang.KeHang_Ctrl"
+            prefHeight="400.0" prefWidth="600.0">
+
+</AnchorPane>
