+++ src/main/resources/com/example/pharmacymanagementsystem_qlht/CN_TimKiem/TKHoaDon/TKHoaDon_GUI.fxml
@@ -19,7 +19,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<Pane fx:id="mainPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="704.0" prefWidth="1200.0" stylesheets="@../css/TimKiemHoaDon.css" xmlns="http://javafx.com/javafx/25" xmlns:fx="http://javafx.com/fxml/1">
+<Pane fx:id="mainPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="704.0" prefWidth="1200.0" stylesheets="@../../css/TimKiemHoaDon.css" xmlns="http://javafx.com/javafx/25" xmlns:fx="http://javafx.com/fxml/1">
     <children>
         <VBox layoutX="10.0" layoutY="14.0">
             <children>
@@ -36,7 +36,7 @@
                         <Label />
                   <ImageView fitHeight="46.0" fitWidth="48.0" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../img/bill-8854.png" />
+                        <Image url="@../../img/bill-8854.png" />
                      </image>
                   </ImageView>
                     </children>
@@ -57,7 +57,7 @@
                   <Region prefHeight="51.0" prefWidth="61.0" />
                   <ImageView fitHeight="20.0" fitWidth="22.0" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../img/time-2623.png" />
+                        <Image url="@../../img/time-2623.png" />
                      </image>
                      <HBox.margin>
                         <Insets right="4.0" top="16.0" />
@@ -91,7 +91,7 @@
                             <graphic>
                                 <ImageView fitHeight="20.0" fitWidth="150.0" pickOnBounds="true" preserveRatio="true">
                                     <image>
-                                        <Image url="@../img/free-search-icon-2911-thumb.png" />
+                                        <Image url="@../../img/free-search-icon-2911-thumb.png" />
                                     </image>
                                     <cursor>
                                         <Cursor fx:constant="DEFAULT" />
@@ -106,7 +106,7 @@
                             <graphic>
                                 <ImageView fitHeight="20.0" fitWidth="34.0" pickOnBounds="true" preserveRatio="true">
                                     <image>
-                                        <Image url="@../img/refresh-3104.png" />
+                                        <Image url="@../../img/refresh-3104.png" />
                                     </image>
                                 </ImageView>
                             </graphic>
