+++ src/main/resources/com/example/pharmacymanagementsystem_qlht/CN_TimKiem/TKKhuyenMai/ChiTietKhuyenMai_GUI.fxml
